pyproject.toml
[tool.poetry]
name = "stoq-server"
version = "2.8.1"
description = "Stoq API backend and syncrhonization daemon and other workers"
authors = ["Stoq Team <dev@stoq.com.br>"]
packages = [{ include = "stoqserver" }]
include = [
    "**/*.mo",
    "bin/*",
    "data/udev/10-stoq.rules",
    "data/sudoers.d/stoqserver",
    "data/supervisor/stoqserver.conf",
    "data/scripts/duplicitybackup.py",
    "data/htsql/config.yml",
    "poetry.lock",
    "setup_old.py",
    "tests/",
    "debian/",
    "README.md",
    "setup.cfg",
    "Makefile",
    "MANIFEST.in",
    "requirements.txt",
]

[[tool.poetry.source]]
name = "gitlab"
url = "https://gitlab.com/api/v4/projects/13882298/packages/pypi/simple"

[tool.poetry.dependencies]
python = "^3.5"

# gitlab packages
stoqnfe = "^1.0.1"
stoqifood = "^1.0.0"
stoqlink = "^1.9"
stoqnonfiscal = "^1.0.0"
stoqlib = "^7.1.0"
stoqsat = "^1.0.1"
stoqpassbook = "^1.0.1"
stoqimendes = "^1.1"
stoqsitef = "^1.22"
stoqconector = "^0.10.1"

# pypi packages
blinker = "^1.3"
Flask = "0.10.1"
Flask-RESTful = "0.3.4"
psycogreen = "1.0.1"
raven = "6.1.0"
requests = "^2.9.1"
stoqdrivers = "^2.0.1"
tzlocal = "1.2.2"
psutil = "3.4.2"
Pygments = "2.1"
gevent = "^20.9.0"
# we need to pin werkzeug its latest version doesn't work for flask < 1
# although flask (0.10.1 used in production) allows it
Werkzeug = "0.10.4"
# this is required by stoq-plugin-nfe
viivakoodi = "^0.8.0"
PyKCS11 = "^1.5.9"
# cffi depends on this but it is not installing correctly
pycparser = "^2.20"
Flask-Log-Request-ID = "^0.10.1"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]
Babel = "^2.8.0"
coverage = "^5.3"
nose = "^1.3.7"
nosexcover = "^1.0.11"
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
pytest-stoq = "^0.7.0"
pycodestyle = "^2.6.0"
pyflakes = "^2.2.0"
bump2version = "^1.0.1"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
